refactor(auth): create auth folder to unify files
